Converted to use device class
The way the app is currently structured means that all classes are
created as inner classes of the Pedalboard class whereas usually
everything would be encapsulated in a module. As the first part of
moving towards a more standard structure, the class-like behaviour of
the Pedalboard has been moved into a Device class making it easier to
change the remaining Pedalboard class to a module.