Converted binary to CLI
Changed to binary to a CLI that can load configurations from a file.
This uses the Mixlib CLI library to parse a file name from the arguments
passed to the binary and creates a device based on this.

An example MainStage configuration has been added based on the start
script.