Added libasound packages to Travis build
UniMIDI requires the libasound and libasound-dev packages to be
installed on Linux machines. Previously, UniMIDI was being required (via
MicroMIDI) in the start script which was not run on the on CI server and
so this wasn't an issue. In the recent reorganisations however, the
MicroMIDI require was moved so that it was required as part of the main
pedalboard code causing the build to fail due to these missing packages.